# Performance optimizations: Remove DI and pagination