--- a/.claude/commands/analyse_app.md
+++ b/.claude/commands/analyse_app.md
@@ -1,0 +1,3 @@
+# Analyse
+
+Review the app logs and if available screenshot and  list the major errors with root cause analysis.