--- a/.claude/commands/app_check.md
+++ b/.claude/commands/app_check.md
@@ -1,2 +1,0 @@
-# check app
-Run on emulator keep in foreground without time outs, wait till run completes then review the logs for errors or opportunities, create a report and propose fixes for my feedback.