--- a/.claude/commands/check_app.md
+++ b/.claude/commands/check_app.md
@@ -1,0 +1,3 @@
+# Check logs and errors
+
+Review the flutter logs and list any errors.