--- a/.claude/commands/screen_shot_app.md
+++ b/.claude/commands/screen_shot_app.md
@@ -1,0 +1,3 @@
+# Take screenshots
+
+Take a screenshot and and list any errors