--- a/free_flight_log_app/lib/data/datasources/database_helper.dart
+++ b/free_flight_log_app/lib/data/datasources/database_helper.dart
@@ -8,6 +8,7 @@ class DatabaseHelper {
   static const _databaseName = "FlightLog.db";
   static const _databaseVersion = 8;
 
+  // Singleton pattern
   DatabaseHelper._privateConstructor();
   static final DatabaseHelper instance = DatabaseHelper._privateConstructor();