--- a/free_flight_log_app/lib/data/repositories/flight_repository.dart
+++ b/free_flight_log_app/lib/data/repositories/flight_repository.dart
@@ -78,26 +78,6 @@ class FlightRepository {
     return maps;
   }
   
-  /// Get paginated flights for better performance
-  Future<List<Flight>> getFlightsPaginated(int offset, int limit) async {
-    LoggingService.debug('FlightRepository: Getting flights (offset: $offset, limit: $limit)');
-    
-    Database db = await _databaseHelper.database;
-    List<Map<String, dynamic>> maps = await db.rawQuery('''
-      SELECT f.*, 
-             ls.name as launch_site_name
-      FROM flights f
-      LEFT JOIN sites ls ON f.launch_site_id = ls.id
-      ORDER BY f.date DESC, f.launch_time DESC
-      LIMIT ? OFFSET ?
-    ''', [limit, offset]);
-    
-    final flights = maps.map((map) => Flight.fromMap(map)).toList();
-    LoggingService.debug('FlightRepository: Retrieved ${flights.length} flights');
-    
-    return flights;
-  }
-
   /// Get total number of flights
   Future<int> getFlightCount() async {
     LoggingService.debug('FlightRepository: Getting flight count');