--- a/free_flight_log_app/lib/presentation/screens/edit_flight_screen.dart
+++ b/free_flight_log_app/lib/presentation/screens/edit_flight_screen.dart
@@ -7,7 +7,6 @@ import '../../data/repositories/flight_repository.dart';
 import '../../data/repositories/site_repository.dart';
 import '../../data/repositories/wing_repository.dart';
 import 'edit_wing_screen.dart';
-import '../../core/dependency_injection.dart';
 
 class EditFlightScreen extends StatefulWidget {
   final Flight flight;
@@ -20,9 +19,9 @@ class EditFlightScreen extends StatefulWidget {
 
 class _EditFlightScreenState extends State<EditFlightScreen> {
   final _formKey = GlobalKey<FormState>();
-  final FlightRepository _flightRepository = serviceLocator<FlightRepository>();
-  final SiteRepository _siteRepository = serviceLocator<SiteRepository>();
-  final WingRepository _wingRepository = serviceLocator<WingRepository>();
+  final FlightRepository _flightRepository = FlightRepository.instance;
+  final SiteRepository _siteRepository = SiteRepository.instance;
+  final WingRepository _wingRepository = WingRepository.instance;
 
   late TextEditingController _notesController;