--- a/free_flight_log_app/lib/presentation/screens/edit_wing_screen.dart
+++ b/free_flight_log_app/lib/presentation/screens/edit_wing_screen.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import '../../data/models/wing.dart';
 import '../../data/repositories/wing_repository.dart';
-import '../../core/dependency_injection.dart';
 
 class EditWingScreen extends StatefulWidget {
   final Wing? wing;
@@ -15,7 +14,7 @@ class EditWingScreen extends StatefulWidget {
 
 class _EditWingScreenState extends State<EditWingScreen> {
   final _formKey = GlobalKey<FormState>();
-  final WingRepository _wingRepository = serviceLocator<WingRepository>();
+  final WingRepository _wingRepository = WingRepository.instance;
 
   late TextEditingController _nameController;
   late TextEditingController _manufacturerController;