--- a/free_flight_log_app/lib/presentation/screens/flight_list_screen.dart
+++ b/free_flight_log_app/lib/presentation/screens/flight_list_screen.dart
@@ -3,6 +3,8 @@ import 'package:provider/provider.dart';
 import '../../data/models/flight.dart';
 import '../../providers/flight_provider.dart';
 import '../../utils/date_time_utils.dart';
+import '../../utils/startup_performance_tracker.dart';
+import '../../services/logging_service.dart';
 import 'add_flight_screen.dart';
 import 'igc_import_screen.dart';
 import 'flight_detail_screen.dart';
@@ -30,9 +32,21 @@ class _FlightListScreenState extends State<FlightListScreen> {
     super.initState();
     // Load flights immediately without delay
     // The splash screen already handled initialization
-    WidgetsBinding.instance.addPostFrameCallback((_) {
+    WidgetsBinding.instance.addPostFrameCallback((_) async {
       if (mounted) {
-        context.read<FlightProvider>().loadFlights();
+        final perfTracker = StartupPerformanceTracker();
+        perfTracker.recordTimestamp('Starting First Data Load');
+        
+        final loadWatch = perfTracker.startMeasurement('First Flight Data Load');
+        await context.read<FlightProvider>().loadFlights();
+        perfTracker.completeMeasurement('First Flight Data Load', loadWatch);
+        
+        perfTracker.recordTimestamp('First Data Load Complete');
+        
+        // Log timing for first meaningful paint
+        final finalReport = perfTracker.generateReport();
+        LoggingService.info('FINAL STARTUP REPORT:\n$finalReport');
+        print('\nFINAL STARTUP REPORT:\n$finalReport');
       }
     });