… .claude/commands/run_app.md
… CLAUDE.md
… free_flight_log_app/lib/data/utils/pagination_result.dart
… free_flight_log_app/lib/services/igc_import_service.dart
… free_flight_log_app/lib/services/site_matching_service.dart
… free_flight_log_app/lib/services/site_migration_service.dart
… free_flight_log_app/lib/utils/startup_performance_tracker.dart
… free_flight_log_app/pubspec.lock
… free_flight_log_app/pubspec.yaml