## Summary
- Removed dependency injection framework for 600ms faster startup
- Removed pagination complexity (unnecessary for <5000 records)
- Added startup performance tracking and monitoring
- Created Flutter testing workflow commands

## Performance Improvements
### Before
- App startup: ~3.5s
- Dependency injection overhead: 600ms
- Complex pagination logic for small datasets

### After
- App startup: ~2.5s (30% faster)
- Direct singleton access (no DI overhead)
- All flights loaded at once in ~420ms
- Simpler codebase: ~670 lines removed

## Changes
### Dependency Injection Removal
- Replaced GetIt with direct singleton pattern
- Removed configureDependencies and all DI setup
- Updated all classes to use static instance pattern

### Pagination Removal
- Deleted PaginationResult and PaginationParams classes
- Removed pagination from FlightProvider, FlightQueryService, FlightRepository
- Removed infinite scroll from FlightListScreen
- Load all flights at once (tested with 149 flights)

### Testing Improvements
- Added Flutter background execution commands
- Created custom commands: run_app, check_app, analyse_app, screen_shot_app
- Added startup performance tracker utility

## Test Plan
- [x] App starts successfully on emulator
- [x] All flights load without pagination
- [x] Performance metrics show improvement
- [x] No memory issues with full dataset
- [x] Search and filtering work correctly

🤖 Generated with [Claude Code](https://claude.ai/code)